修复订阅管理功能：修复JSON字段名不匹配、添加缺失的后端模型字段、修复API路径、改进日期显示

+++ backend/internal/service/scaler_service.go
@@ -363,8 +363,8 @@ func (s *scalerService) GetGroupMetrics(groupID string) (*GroupMetrics, error) {
 		totalCPU += server.CPUUsage
 		totalMemory += server.MemoryUsage
 		metrics.TotalConnections += server.Connections
-		metrics.TotalBandwidthUp += server.BandwidthUp
-		metrics.TotalBandwidthDown += server.BandwidthDown
+		metrics.TotalBandwidthUp += int64(server.BandwidthUp)
+		metrics.TotalBandwidthDown += int64(server.BandwidthDown)
 	}
 
 	metrics.ServerCount = len(servers)
